engine/docker: Use current dir's root dir as mount point
The mount point is changed to current dir's root dir by recursively finding
its parent directory until the parent of the parent directory is the root,
i.e., "/". In the finding process, ubuild try to resolve the actual path
that is pointed to by its symbolic link.

Assuming that the current dir is "~/developer/project", which is a symbolic
link pointing to "/mnt/disk/project".  The finding process is:

- Resolve symbolic link: "/mnt/disk/project" -> "/mnt/disk/project".
- Go to "/mnt/disk", since the parent of parent dir is "/mnt", continue;
- Go to "/mnt", since the parent of parent dir is "/", it meets the
  stopping condition.

The mount point will be "/mnt".

Signed-off-by: Xuewei Niu <niuxuewei.nxw@antgroup.com>